Sensor/requirements.txt
grpcio>=1.53.0
grpcio-tools>=1.53.0
protobuf>=3.20.0